controller/src/requirements.txt
Django==2.0.4
djangorestframework==3.8.2
pytz==2018.4
PyYAML==3.12
ua-parser==0.8.0
user-agents==1.1.0
